add form m state - form m singleton will retrieve its issues when the singleton is initialized.
The plan is not to have the issues attached to the form M object but retrieved async after the form M has been initialized, the way the bid is done. This will eventually lead to a better separation of concerns between the directives with the form M singleton as a single source of truth.