form m bid - experimented wih setting bid form on singleton form M object.
It did not work. Well not exactly - if we are visiting the form M detail page from another tab, the form is set correctly. However if we destroy and recreate the form M singleton e.g by asking angular ui router to redraw the state after form M successfully saved, the form would have been destroyed and there is no other way to recover the form except by leaving that tab for another tab and then coming back to it - and this I don't know how to do.