treasury allocation app - dismissing editing allocation should be a function of edit allocation directive.
initially, the search allocation directive provides its own ui element for dismissing the edit allocation directive. But this is fragile because edit allocation directive might have changed the allocation in such a way that it will leave the allocation in an inconsistent state. When a user requires dismissal (changes her mind and does not wish to edit again) then edit allocation directive should be able to reset the allocation.